+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   * Skeleton for generating and consuming `refreshTokens`
   * Decorators for injecting the `attribute` and `resource` into a subclass's method definition
   * keys and values in `Actions` enum relating to `OrderEvent`
+  * update `Actions.PURCHASE`, `Actions.SHIP` and `Actions.RETURN` values to be uppercase
 ### Removed
   * `moment` is no longer a peer dependency, just a `devDependency`
 ### Changed